Updated files

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use PDF;
+use Illuminate\Support\Facades\Auth;
 // use PDF;
 
 class ProductController extends Controller
@@ -14,12 +15,17 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $product = Product::orderBy('created_at', 'DESC')->get();
+        // echo "jalan";
+        // die();
+        $product = Product::orderBy('created_at', 'DESC')->paginate(5);
 
         return view('products.index', compact('product'));
     }
     public function cetakPdf()
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('product')->with('error', 'You do not have permission to Print products.'); 
+        }
         $product = Product::get()->all();
 
         $pdf = PDF::loadView('cetakPdf', ['product' => $product]);
@@ -31,6 +37,9 @@ class ProductController extends Controller
      */
     public function create()
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('product')->with('error', 'You do not have permission to create products.'); 
+        }
         // $imageName = time().'.'.request()->file('image')->extension();
         // request()->image->move(public_path('images'), $imageName);
         return view('products.create');
@@ -41,6 +50,9 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('products')->with('error', 'You do not have permission to store products.'); 
+        }
         $request->validate([
             'title' => 'required',
             'price' => 'required',
@@ -64,8 +76,7 @@ class ProductController extends Controller
         $product->tahun_pengadaan = $request->tahun_pengadaan;
 
         $product->save();
-        return back()->withSuccess('Product Successfully Created!!!!');
-    }
+        return redirect()->route('product')->with('success', 'Product Successfully Created!');    }
 
 
     /**
@@ -73,16 +84,32 @@ class ProductController extends Controller
      */
     public function show(string $id)
     {
+        
         $product = Product::findOrFail($id);
 
         return view('products.show', compact('product'));
     }
+    public function search(Request $request)
+{
+    $query = $request->input('query'); // Ambil input dari user
+    $product = Product::where('title', 'LIKE', "%$query%") // Ganti 'title' dengan kolom yang sesuai
+                      ->orWhere('description', 'LIKE', "%$query%")
+                      ->orWhere('tahun_pengadaan', 'LIKE', "%$query%")
+                      ->orWhere('lokasi_pengguna', 'LIKE', "%$query%")
+                      ->paginate(5);
+
+    return view('products.index', compact('product'));
+}
+
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(string $id)
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('product')->with('error', 'You do not have permission to edit products.'); 
+        }
         $product = Product::findOrFail($id);
 
         return view('products.edit', compact('product'));
@@ -93,6 +120,9 @@ class ProductController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('product')->with('error', 'You do not have permission to Update products.'); 
+        }
         $product = Product::findOrFail($id);
 
         $product->update($request->all());
@@ -105,6 +135,9 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+        if (Auth::user()->type != 'admin') { 
+            return redirect()->route('product')->with('error', 'You do not have permission to delete products.'); 
+        }
         $product = Product::findOrFail($id);
 
         $product->delete();
